FLUID-4661: beginning to defer more to mejs
Added onFullScreen and onExitFullScreen events as well as functions for requesting and cancelling fullscreen, to VideoPlayer_media.js. These make use of functionality provided by MediaElement.js.

Currently I've only modified the code which used the fullscreen events. Need to use the new request and cancel functions as well in a future commit.