FLUID-4661: refactored fullscreen method
Refactored the fullscreen method into a toggleFullscreen method that only changes the model. The media component now listens to a model changed event for the fullscreen property and performs the necessary action on the video player.

Was able to remove one more name finding utility, but need to find a new way of handling the supportsFullScreen context in order to remove the fullscreen function name code.